requirements.txt
Jinja2==2.7.3
MarkupSafe==0.23
PyYAML==3.11
backports.ssl-match-hostname==3.4.0.2
docker-py==0.7.0
gnureadline==6.3.3
requests==2.5.0
six==1.8.0
websocket-client==0.23.0
wsgiref==0.1.2
